Revert ":art: edit: 웹소켓 서비스 정의"
This reverts commit 29bf06d4ac7f55288f0f898ca29f5bddca884d02.
