+++ src/main/java/com/wepong/pongdang/service/GameRoomService.java
@@ -1,7 +1,9 @@
 package com.wepong.pongdang.service;
 
 import com.wepong.pongdang.dto.request.GameRoomRequestDTO;
+import com.wepong.pongdang.dto.response.ChatResponseDTO;
 import com.wepong.pongdang.dto.response.GameRoomResponseDTO;
+import com.wepong.pongdang.dto.response.WebSocketResponseDTO;
 import com.wepong.pongdang.entity.GameEntity;
 import com.wepong.pongdang.entity.GameLevelEntity;
 import com.wepong.pongdang.entity.GameRoomEntity;
@@ -14,6 +16,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,6 +29,7 @@ public class GameRoomService {
 
 	private final PlayerDAO playerDAO;
 	private final GameRoomRepository gameRoomRepository;
+	private final SimpMessagingTemplate messagingTemplate;
 	private final AuthService authService;
 	private final GameService gameService;
 	private final GameLevelService gameLevelService;
@@ -106,5 +110,49 @@ public class GameRoomService {
 		room.updateHost(userEntity);
 
 		gameRoomRepository.save(room);
+	}
+
+	public void sendRoom(Long roomId, String type, Object data) {
+		WebSocketResponseDTO payload = WebSocketResponseDTO.builder()
+				.type(type)
+				.data(data)
+				.build();
+
+		messagingTemplate.convertAndSend("/topic/gameroom/" + roomId, payload);
+	}
+
+	public void sendList(String type, Object data) {
+		WebSocketResponseDTO payload = WebSocketResponseDTO.builder()
+				.type(type)
+				.data(data)
+				.build();
+
+		messagingTemplate.convertAndSend("/topic/gameroom", payload);
+	}
+
+	public void sendGame(Long roomId, String type, Object data) {
+		WebSocketResponseDTO payload = WebSocketResponseDTO.builder()
+				.type(type)
+				.data(data)
+				.build();
+
+		messagingTemplate.convertAndSend("/topic/game" + roomId, payload);
+	}
+
+	public void sendGame(Long roomId, String type) {
+		WebSocketResponseDTO payload = WebSocketResponseDTO.builder()
+				.type(type)
+				.build();
+
+		messagingTemplate.convertAndSend("/topic/game" + roomId, payload);
+	}
+
+	public void sendTest(String type, ChatResponseDTO chat) {
+		WebSocketResponseDTO payload = WebSocketResponseDTO.builder()
+				.type(type)
+				.data(chat)
+				.build();
+
+		messagingTemplate.convertAndSend("/topic/test", payload);
 	}
 }
