+++ src/main/java/com/wepong/pongdang/socket/StompEventListener.java
@@ -6,7 +6,6 @@ import com.wepong.pongdang.entity.enums.GameRoomStatus;
 import com.wepong.pongdang.model.multi.turtle.PlayerService;
 import com.wepong.pongdang.model.multi.turtle.TurtleGameService;
 import com.wepong.pongdang.service.GameRoomService;
-import com.wepong.pongdang.service.WebSocketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
@@ -27,7 +26,6 @@ public class StompEventListener {
     private final GameRoomService gameRoomService;
     private final PlayerService playerService;
     private final TurtleGameService turtleGameService;
-    private final WebSocketService webSocketService;
 
     private final Map<Long, List<TurtlePlayerDTO>> gameStartPlayersMap = new ConcurrentHashMap<>();
 
@@ -50,8 +48,8 @@ public class StompEventListener {
             accessor.getSessionAttributes().put("nickname", player.getNickname());
 
             List<TurtlePlayerDTO> players = playerService.getPlayers(roomId);
-            webSocketService.sendRoom(roomId, "enter", players);
-            webSocketService.sendList(gameRoomService.selectAll());
+            gameRoomService.sendRoom(roomId, "enter", players);
+            gameRoomService.sendList("list", gameRoomService.selectAll());
 
         } else if(accessor.getSessionAttributes().get("type").equals("game")) {
             // (1) 게임 시작 전이면 검증 건너뛰고, 그냥 세션 등록
@@ -73,11 +71,11 @@ public class StompEventListener {
                 Map<String, Object> msg = new HashMap<>();
                 msg.put("reason", "no_player_info");
                 msg.put("targetUrl", "/gameroom");
-                webSocketService.sendGame(roomId, "force_exit", msg);
+                gameRoomService.sendGame(roomId, "force_exit", msg);
                 try { Thread.sleep(50); } catch (InterruptedException ignored) {}
             }
         } else if(accessor.getSessionAttributes().get("type").equals("list")) {
-            webSocketService.sendList(gameRoomService.selectAll());
+            gameRoomService.sendList("list", gameRoomService.selectAll());
         }
     }
 
@@ -108,11 +106,11 @@ public class StompEventListener {
 
                     } else {
                         gameRoomService.deleteRoom(roomId);
-                        webSocketService.sendList(gameRoomService.selectAll());
+                        gameRoomService.sendList("delete", gameRoomService.selectAll());
                     }
                 }
 
-                webSocketService.sendRoom(roomId, "exit", players);
+                gameRoomService.sendRoom(roomId, "exit", players);
             }
         } else if("game".equals(type)) {
             if(!status.equals(GameRoomStatus.WAITING)) {
@@ -128,7 +126,7 @@ public class StompEventListener {
 
                     } else {
                         turtleGameService.removeGame(roomId);
-                        webSocketService.sendList(gameRoomService.selectAll());
+                        gameRoomService.sendList("delete", gameRoomService.selectAll());
                     }
                 }
 
@@ -136,10 +134,10 @@ public class StompEventListener {
                     Map<String, Object> msg = new HashMap<>();
                     msg.put("reason", "connection_error");
                     msg.put("targetUrl", "/gameroom");
-                    webSocketService.sendGame(roomId, "force_exit", msg);
+                    gameRoomService.sendGame(roomId, "force_exit", msg);
                 }
             }
         }
-        webSocketService.sendList(gameRoomService.selectAll());
+        gameRoomService.sendList("list", gameRoomService.selectAll());
     }
 }
